+++ doc/source/changelog.rst
@@ -66,6 +66,7 @@ Command line
 * ``qitoolchain import-package`` learned ``--batch``
 * ``qitoolchain import-package`` learned to import package directory
 * :cmake:function:`qi_add_test` can now handle test script as target instead of binary.
+* ``qibuild make`` learned ``--verbose-make``
 
 CMake
 ++++++